Use UUID for request rather than a constructed one
The constructed request ID isn't guaranteed to be unique across
controllers. This probably won't be an issue, but if the ID is used to
tag resources to ensure they can be cleaned up, there is a small
potential for conflict if multiple controllers are in use. It's safer
and easier to just generate a UUID instead.

In the name of simplicity, this also makes the name and ID into regular
fields, as well.