Allow non-leaders to see responses
For whatever reason, non-leaders are blocked from reading app-level
relation data set by the leader. This means they can't read the list of
requests nor compare the hashes of those requests to see if the
responses are complete. This works around that by just having
non-leaders assume that any provided response is complete.